edit

+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/model/entity/ProductEntity.java
@@ -40,7 +40,7 @@ public class ProductEntity {
     @UpdateTimestamp
     private LocalDateTime updateDate;
 
-    private Float discount;
+    private long discount;
 
     @Column(name = "description")
     private String description;
